requirements.txt
Django==5.2.5
djangorestframework==3.16.0
pandas==2.3.2
Requests==2.32.5
Scrapy==2.11.2
SQLAlchemy==2.0.34
